requirements.txt
Pillow==9.0.1
tkinterdnd2~=0.3.0
darkdetect~=0.7.1